Change TimeId counter types
Negative slabs will be used for initialization.

Substeps should never be negative, but were changed to a fixed-width
type for consistency.

+++ src/Time/TimeId.hpp
@@ -7,6 +7,7 @@
 #pragma once
 
 #include <cstddef>
+#include <cstdint>
 #include <functional>
 #include <iosfwd>
 
@@ -27,7 +28,7 @@ class TimeId {
   /// Create a TimeId at the start of a step.  If that step is at the
   /// (evolution-defined) end of the slab the TimeId will be advanced
   /// to the next slab.
-  TimeId(const bool time_runs_forward, const size_t slab_number,
+  TimeId(const bool time_runs_forward, const int64_t slab_number,
          const Time& time) noexcept
       : time_runs_forward_(time_runs_forward),
         slab_number_(slab_number),
@@ -38,8 +39,9 @@ class TimeId {
   }
   /// Create a TimeId at a substep at time `time` in a step starting
   /// at time `step_time`.
-  TimeId(const bool time_runs_forward, const size_t slab_number,
-         const Time& step_time, const size_t substep, const Time& time) noexcept
+  TimeId(const bool time_runs_forward, const int64_t slab_number,
+         const Time& step_time, const uint64_t substep,
+         const Time& time) noexcept
       : time_runs_forward_(time_runs_forward),
         slab_number_(slab_number),
         step_time_(step_time),
@@ -51,10 +53,10 @@ class TimeId {
   }
 
   bool time_runs_forward() const noexcept { return time_runs_forward_; }
-  size_t slab_number() const noexcept { return slab_number_; }
+  int64_t slab_number() const noexcept { return slab_number_; }
   /// Time at the start of the current step
   const Time& step_time() const noexcept { return step_time_; }
-  size_t substep() const noexcept { return substep_; }
+  uint64_t substep() const noexcept { return substep_; }
   /// Time of the current substep
   const Time& time() const noexcept { return time_; }
 
@@ -69,9 +71,9 @@ class TimeId {
   void canonicalize() noexcept;
 
   bool time_runs_forward_{false};
-  size_t slab_number_{0};
+  int64_t slab_number_{0};
   Time step_time_{};
-  size_t substep_{0};
+  uint64_t substep_{0};
   Time time_{};
 };
